repo and service testing done

+++ src/main/java/com/github/roknikolic/springcrudapp/product/ProductService.java
@@ -10,7 +10,7 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
-    public Iterable<Product> readAll() {
+    public Iterable<Product> read() {
         return productRepository.findAll();
     }
 
@@ -18,16 +18,31 @@ public class ProductService {
         return productRepository.findById(id).orElse(null);
     }
 
-    public void remove(Integer id) {
-        productRepository.deleteById(id);
+    public Product remove(Integer id) {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if (optionalProduct.isPresent()) {
+            productRepository.deleteById(id);
+            return optionalProduct.get();
+        } else {
+            return null;
+        }
     }
 
     public Product create(Product product) {
-        product.setId(null);
+        if ((product.getName() == null) ||  (product.getDescription() == null) || (product.getPrice() == null)) {
+            return null;
+        }
+        product.setId(null); // We force auto generated IDs
         return productRepository.save(product);
     }
 
     public Product update(Product product) {
+        if ((product.getName() == null) ||
+            (product.getDescription() == null) ||
+            (product.getPrice() == null) ||
+            (product.getId() == null)) {
+            return null;
+        }
         Optional<Product> optionalProduct = productRepository.findById(product.getId());
         if (optionalProduct.isPresent()) {
             productRepository.save(product);
